Add support for non-standard path
A module written on C requires tarantool header files. Usually a
rockspec checks for its existence and pass a directory to use inside a
build system (say, cmake).

The check for existence may be performed like so:

 | external_dependencies = {
 |     TARANTOOL = {
 |         header = 'tarantool/module.h'
 |     }
 | }

While the directory may be passed in this way:

 | build = {
 |     type = 'cmake',
 |     variables = {
 |         TARANTOOL_DIR='$(TARANTOOL_DIR)',
 |     }
 | }

Consider 'ckit' branch of 'tarantool/modulekit' repository for a full
example.

The check for external dependencies fails when tarantool is installed
outside of '/usr/ or '/usr/local/' directories. Consider [1] for more
information about the problem.

To fix the problem hardened.PREFIX, which is filled during tarantool
build with CMAKE_INSTALL_PREFIX value, is added to the default value of
'external_deps_dirs' list of directories. Luarocks use this value to
resolve external dependencies.

The patch is based on the following patch for luarocks-2:

- b2dde9a4 ('[luarocks] Add support for non-standard path')

[1]: https://github.com/tarantool/tarantool/issues/3148